Allow for an additional 'scope_change' breakpoint type.
This breakpoint is meant to signal to the debugger engine whether it should
break on changes in the active scope. This breakpoint will trigger when
the active scope of the current state of the debugger engine changes.

This, combined with another PR regarding the validity of configured
breakpoints should allow for the detection of breakpoints on lines of code
that are not valid.