A new flag to breakpoints to show whether it was resolved.
A resolved breakpoint is one that the debugging engine knows for sure is valid.
This means that either the exception name, function name, or file/line number
combination is valid to break on.

This new flag would be present for breakpoint_set and breakpoint_list command
results if the engine supports the notion of resolved breakpoints.